revert temp rename
